Abstract out gift_aid_section to its own file.
