+++ src/user/components/my_details/gift_aid_buttons.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 export default ({ confirmation, which_text, which_delete, reset }) =>
   <div>
-    <p><b>You have made a Gift Aid Declaration.</b></p>
+    <h4>You have made a Gift Aid Declaration</h4>
     <p>
       <i>
         If you no longer pay an amount of UK Income Tax and/or Capital Gains Tax
