+++ src/user/components/my_details/membership_details.js
@@ -6,6 +6,7 @@ import EndMembershipForm from './end_membership_form.js'
 import EditDetails from './edit_details.js'
 import ConfirmGiftAid from '../../../shared/components/confirm_deletion.js'
 import GiftAidButtons from './gift_aid_buttons.js'
+import GiftAidSection from './gift_aid_section.js'
 
 export default (props) =>
   <div>
@@ -28,38 +29,6 @@ const EndMembershipLoader = (props) =>
     fields={['activation_status', 'notes', 'user_notes', 'deletion_date']}
   />
 
-const GiftAidSection = ({toggle_gift_aid}) => {
-  const handleSubmit = e => {
-    e.preventDefault()
-    toggle_gift_aid()
-  }
-  return (
-    <div>
-      <p><b>You have not made a Gift Aid Declaration</b></p>
-      <p>
-        <i>
-          You can boost the value of your subscription/donation to the Friends by 25p of
-          Gift Aid for every £1 subscribed/donated. Gift Aid is reclaimed by the Friends
-          from the tax you pay for the current tax year. If you wish to Gift Aid your
-          subscription/donations to the Friends for this year, future years and the past
-          4 years please tick the box below.
-        </i>
-      </p>
-      <form onSubmit={handleSubmit}>
-        <div>
-          <input type='checkbox' required/>
-          <p>
-            I wish to make a Gift Aid Declaration to the Friends. I am a UK taxpayer and
-            understand that if I pay less Income Tax and/or Capital Gains Tax than the
-            amount of Gift Aid claimed on all my donations in that tax year it is my
-            responsibility to pay any difference.
-          </p>
-        </div>
-        <button type='submit'>Confirm Gift Aid</button>
-      </form>
-    </div>
-  )
-}
 
 const EndMembership = reduxForm(
   { form: 'member'
